api/requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Data processing
pyyaml==6.0.1
pandas==2.1.3

# Rate limiting and middleware
slowapi==0.1.9

# Environment configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Async file operations
aiofiles==24.1.0

# Response compression
python-multipart==0.0.6

# Data export formats
openpyxl==3.1.2
pyarrow==14.0.1

# Optional dependencies (for visualization if needed)
matplotlib==3.8.2

# For testing the API
requests==2.31.0

# Metrics and monitoring
prometheus-client==0.19.0
psutil==5.9.6
